// packages/openapi-bench/generators/openapi-generator/client/model/problem.ts
/**
 * OpenAPI Bench API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Problem { 
    type: string;
    title: string;
    status: number;
    detail?: string;
}