// packages/openapi-bench/generators/openapi-generator/client/api/pets.service.ts
export interface CreatePetRequestParams {
    pet: Pet;
}