// packages/openapi-bench/generators/openapi-generator/client/api/params.service.ts
export interface HeaderAndCookieParamsRequestParams {
    xRequestId: string;
    xTraceFlags?: Array<string>;
    xRetryCount?: number;
    session?: string;
}

export interface MultiPathParamsRequestParams {
    stringId: string;
    intId: number;
    uuidId: string;
    enumId: PetStatus;
}

export interface QueryStylesRequestParams {
    required: string;
    tagsExplode?: Array<string>;
    tagsCsv?: Array<string>;
    tagsPipe?: Array<string>;
    tagsSpace?: Array<number>;
    filter?: PetFilter;
    point?: QueryStylesPointParameter;
    since?: string;
    day?: string;
    flag?: boolean;
    sort?: 'asc' | 'desc';
}