// packages/openapi-bench/generators/openapi-generator/client/api/bodies.service.ts
export interface SubmitFormRequestParams {
    username: string;
    password: string;
    remember?: boolean;
    scopes?: Array<string>;
}

export interface UploadBinaryRequestParams {
    body: Blob;
}

export interface UploadFilesRequestParams {
    title: string;
    file: Blob;
    attachments?: Array<Blob>;
    meta?: UploadMeta;
}