// packages/openapi-bench/generators/openapi-generator/client/model/pet.ts
/**
 * OpenAPI Bench API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NullableColor } from './nullable-color';
import { Category } from './category';
import { Priority } from './priority';
import { Tag } from './tag';
import { PetStatus } from './pet-status';


export interface Pet { 
    readonly id: number;
    name: string;
    status: PetStatus;
    priority?: Priority;
    color?: NullableColor | null;
    nickname?: string | null;
    photoUrls: Array<string>;
    tags?: Array<Tag>;
    category?: Category;
    birthDate?: string;
    vaccinated?: boolean;
    weightKg?: number;
    readonly createdAt?: string;
    secretChipCode?: string;
    /**
     * Veraltet, wird entfernt
     * @deprecated
     */
    legacyCode?: string;
    attributes?: { [key: string]: string; };
    metadata?: { [key: string]: any; };
}
export namespace Pet {
}