// packages/openapi-bench/generators/openapi-generator/client/model/pet-page.ts
/**
 * OpenAPI Bench API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Pet } from './pet';


export interface PetPage { 
    total: number;
    limit: number;
    offset: number;
    items: Array<Pet>;
}