// packages/openapi-bench/generators/openapi-generator/client/model/pet-filter.ts
/**
 * OpenAPI Bench API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PetStatus } from './pet-status';


export interface PetFilter { 
    name?: string;
    status?: PetStatus;
    minAge?: number;
}
export namespace PetFilter {
}